feat: fixed Alert queries for all types of queries

+++ pkg/alerts/alertsHandler/alertsHandler.go
@@ -697,7 +697,7 @@ func convertToSiglensAlert(lmDetails alertutils.LogLinesFile) []*alertutils.Mini
 			LogTextHash:     entry.LogTextHash,
 			LogLevel:        entry.LogLevel,
 			Condition:       alertutils.IsAbove,
-			Value:           float32(entry.Value),
+			Value:           float64(entry.Value),
 			EvalFor:         0,
 			EvalInterval:    1,
 			Message:         "Minion search " + alert_name,
